change to openjdk11 and openjfx11

+++ BaseQspEngine/src/main/java/com/qsp/player/libqsp/service/AudioPlayer.java
@@ -14,7 +14,7 @@ public class AudioPlayer {
 
     public void playFile(GameStatus gameStatus, String path, int volume) {
         closeAllFiles();
-        audioInterface = audioFactory.createNewPlayer(gameStatus, path,Double.parseDouble(String.valueOf(volume)));
+        audioInterface = audioFactory.createNewPlayer(gameStatus, path,Float.parseFloat(String.valueOf(volume)));
         audioInterface.setVolume(volume);
         audioInterface.start();
     }
